Update to booking form

+++ app/Http/Controllers/LocationController.php
@@ -8,19 +8,19 @@ use Illuminate\Http\Request;
 
 class LocationController extends Controller
 {
-        /**
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
-        return Location::with('media')->get()->map(function($q){
+        return Location::with('media')->get()->map(function ($q) {
             return [
                 'id' => $q->id,
                 'facility_name' => $q->facility_name,
                 'name' => $q->name,
-                'image' => ($q->media->last()) ? $q->media->last()->getFullUrl() : ''
+                'image' => ($q->media->last()) ? $q->media->last()->getUrl() : ''
             ];
         });
     }
@@ -33,7 +33,7 @@ class LocationController extends Controller
     public function create(Request $request)
     {
         $location = Location::create([
-            'customer_id'=> Customer::current()->id,
+            'customer_id' => Customer::current()->id,
             'name' => $request->name,
             'facility_name' => $request->facility_name,
         ]);
@@ -53,6 +53,11 @@ class LocationController extends Controller
 
     public function delete(Location $location)
     {
+        // check if a location has units:
+        if ($location->units->count() !== 0) {
+            return ["success" => false, "error" => "location_has_units"];
+        }
         $location->delete();
+        return ["success" => true];
     }
 }
